# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "flow-talk-server 接口",
    "description": "当前服务已实现用户认证、会话、消息、WebSocket 实时通信、设备、群管理、消息回执、在线状态和消息搜索接口。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/api/conversations/{conversation_id}/messages": {
      "post": {
        "tags": [
          "消息"
        ],
        "summary": "发送消息",
        "description": "当前用户向指定会话发送消息，服务端写入 messages 并更新会话最后消息。",
        "operationId": "sendMessage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              },
              "example": {
                "client_msg_id": "client-001",
                "message_type": "text",
                "content": {
                  "text": "hello"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "消息"
        ],
        "summary": "获取历史消息",
        "description": "分页拉取指定会话的历史消息。",
        "operationId": "listMessages",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "before_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePageApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/conversations/{conversation_id}/read": {
      "post": {
        "tags": [
          "消息"
        ],
        "summary": "标记会话已读",
        "description": "更新当前用户在指定会话中的 last_read_message_id。",
        "operationId": "markConversationRead",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              },
              "example": {
                "last_read_message_id": 1001
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "标记成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStateApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录或注册接口返回的 token。"
      }
    },
    "responses": {
      "ValidationError": {
        "description": "参数校验失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorApiResponse"
            },
            "example": {
              "code": 400,
              "data": null,
              "message": "参数校验失败"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未登录或登录已失效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorApiResponse"
            },
            "example": {
              "code": 401,
              "data": null,
              "message": "未登录或登录已失效"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "无权访问目标资源",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorApiResponse"
            },
            "example": {
              "code": 403,
              "data": null,
              "message": "无权访问该会话"
            }
          }
        }
      },
      "NotFound": {
        "description": "目标资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorApiResponse"
            },
            "example": {
              "code": 404,
              "data": null,
              "message": "会话不存在"
            }
          }
        }
      },
      "InternalError": {
        "description": "服务器内部错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorApiResponse"
            },
            "example": {
              "code": 500,
              "data": null,
              "message": "服务器内部错误"
            }
          }
        }
      }
    },
    "schemas": {
...
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "client_msg_id",
          "message_type",
          "content"
        ],
        "properties": {
          "client_msg_id": {
            "type": "string",
            "description": "客户端生成的消息 ID，用于幂等去重",
            "example": "client-001"
          },
          "message_type": {
            "type": "string",
            "description": "消息类型",
            "enum": [
              "text",
              "image",
              "file"
            ],
            "example": "text"
          },
          "content": {
            "type": "object",
            "description": "消息内容，不同消息类型使用不同结构",
            "example": {
              "text": "hello"
            }
          }
        }
      },
      "MarkReadRequest": {
        "type": "object",
        "required": [
          "last_read_message_id"
        ],
        "properties": {
          "last_read_message_id": {
            "type": "integer",
            "format": "int64",
            "example": 1001
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1001
          },
          "conversation_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "sender_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "client_msg_id": {
            "type": "string",
            "example": "client-001"
          },
          "message_type": {
            "type": "string",
            "example": "text"
          },
          "content": {
            "type": "object",
            "example": {
              "text": "hello"
            }
          },
          "status": {
            "type": "string",
            "example": "normal"
          },
          "sent_at": {
            "type": "string",
            "example": "2026-06-29T12:00:00+08:00"
          }
        }
      },
      "MessagePage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "next_before_id": {
            "type": "integer",
            "format": "int64",
            "example": 1000
          },
          "has_more": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "ReadState": {
        "type": "object",
        "properties": {
          "conversation_id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "last_read_message_id": {
            "type": "integer",
            "format": "int64",
            "example": 1001
          },
          "last_read_at": {
            "type": "string",
            "example": "2026-06-29T12:05:00+08:00"
          }
        }
      },
      "MessageApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/Message"
          },
          "message": {
            "type": "string",
            "example": "发送消息成功"
          }
        }
      },
      "MessagePageApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/MessagePage"
          },
          "message": {
            "type": "string",
            "example": "获取历史消息成功"
          }
        }
      },
      "ReadStateApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ReadState"
          },
          "message": {
            "type": "string",
            "example": "标记已读成功"
          }
        }
      },
      "ErrorApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 400
          },
          "data": {
            "nullable": true,
            "example": null
          },
          "message": {
            "type": "string",
            "description": "接口信息或错误信息",
            "example": "参数校验失败"
          }
        }
      },
...
    }
  }
}